backend/requirements.txt
# Core Flask dependencies
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-CORS>=4.0.0
Flask-JWT-Extended>=4.6.0
Flask-Migrate>=4.0.5

# Database drivers
SQLAlchemy>=2.0.25
psycopg2-binary>=2.9.9
PyMySQL>=1.1.0
pymongo>=4.6.0
# Removed cx-Oracle and cassandra-driver for Python 3.13 compatibility

# Document processing
PyMuPDF>=1.23.8
python-docx>=1.1.0
openpyxl>=3.1.2
pandas>=2.1.4
# Removed unstructured[all-docs] for Python 3.13 compatibility

# Code parsing (optional - remove if causing issues)
tree-sitter>=0.20.0
# Removed specific tree-sitter language bindings for Python 3.13 compatibility

# Vector storage and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
numpy>=1.24.4

# LLM integration
google-generativeai>=0.3.2
openai>=1.6.1

# OAuth and authentication
authlib>=1.2.1
requests>=2.31.0
requests-oauthlib>=1.3.1

# Utilities
python-dotenv>=1.0.0
celery>=5.3.4
redis>=5.0.1
gunicorn>=21.2.0
python-multipart>=0.0.6

# Development
pytest>=7.4.3
pytest-flask>=1.3.0
black>=23.11.0
flake8>=6.1.0

# Production server
gunicorn>=21.2.0
gevent>=23.7.0

# Monitoring
prometheus-client>=0.19.0

# Performance monitoring
psutil>=5.9.0

# Additional dependencies for new features
werkzeug>=3.0.0
click>=8.1.0
itsdangerous>=2.1.0
jinja2>=3.1.0
markupsafe>=2.1.0
